Create separate property class for Derived Fields
